hooks moved, add dist to gitignore
